move parsing functions into single file - add parse function using zod
