+++ src/lib/util.ts
@@ -1,7 +1,3 @@
-import { DbUser } from "./db"
-
-export const base64Encode = (str: string) => Buffer.from(str).toString("base64")
-
 export const formatDate = (date: Date) =>
   `${date.getDate()}/${date.getMonth() + 1}/${date.getFullYear()}`
 export const formatDateHtml = (date: Date) =>
@@ -28,82 +24,8 @@ export const multipleClasses = (...args: (string | undefined)[]) =>
     return `${prev} ${curr}`
   }, "")
 
-export async function fetchJsonData<T = any>(url: string, accessToken: string): Promise<T> {
-  const response = await fetch(url, {
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      Accept: "application/json",
-    },
-  })
-  const report: T = await response.json()
-
-  if (!response.ok) {
-    throw { ...report, url }
-  }
-
-  return report
-}
-
-export async function postJsonData<T = any>(url: string, json: any): Promise<T> {
-  const response = await fetch(url, {
-    method: "POST",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(json),
-  })
-  const report: T = await response.json()
-
-  if (!response.ok) {
-    throw { ...report, url }
-  }
-
-  return report
-}
-
 export type DeepPartial<T> = T extends object
   ? {
       [P in keyof T]?: DeepPartial<T[P]>
     }
   : T
-
-export function alreadyFilledIn(doc: FirebaseFirestore.DocumentSnapshot<DbUser>): {
-  items: boolean
-  doneeDetails: boolean
-} {
-  const dbData = doc.data()
-
-  if (!dbData) return { items: false, doneeDetails: false }
-
-  return {
-    items: Boolean(dbData.items) && Boolean(dbData.date),
-    doneeDetails: Boolean(
-      dbData.donee &&
-        dbData.donee.companyAddress &&
-        dbData.donee.companyName &&
-        dbData.donee.country &&
-        dbData.donee.largeLogo &&
-        dbData.donee.registrationNumber &&
-        dbData.donee.signatoryName &&
-        dbData.donee.signature &&
-        dbData.donee.smallLogo
-    ),
-  }
-}
-
-export function isJpegOrPngDataURL(str: string): boolean {
-  if (!str.startsWith("data:image/jpeg;base64,") && !str.startsWith("data:image/png;base64,")) {
-    return false
-  }
-  const regex = /^data:image\/(jpeg|png);base64,([a-zA-Z0-9+/]*={0,2})$/
-  return regex.test(str)
-}
-
-export const toBase64 = (file: File) =>
-  new Promise<string>((resolve, reject) => {
-    const reader = new FileReader()
-    reader.readAsDataURL(file)
-    reader.onload = () => resolve(reader.result as string)
-    reader.onerror = reject
-  })
